Initial Commit
Initial commit for the Rightmove House Prices scraper. 
Note, this script currently only works for prices between £1000, and £1,000,000. The script can be easily adapted to work for other cases such as price >£1,000,000, and I am working on allowing it to dynamically allow different price ranges. 
Also working on scraping data on square footage.